:truck: move getAssessment usecase to `src/shared/`
